requirements.txt
vk==2.0.2
PySimpleGUI
Pillow
